ConfigTraitを追加し、cfFormMailerクラスに設定管理機能を統合
